// src/app/models/company.ts
export class Company{
    companyId: number;
    companyName: string;
    email: string;
    fax: string;
    contactNo: string;
    phoneNo: string;
    addressOne: string;
    addressTwo: string;
    city:string;
    postalCode: string;
    state: string;
    country: string;
    status: string;
    companyCode: string;
    address: string;

    constructor(){
        
    }
}